feature: changed to use threeds not coremethods data
