+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/data/remote/service/CoreMethodsService.kt
@@ -1,7 +1,6 @@
 package com.mercadopago.sdk.android.coremethods.data.remote.service
 
 import com.mercadopago.sdk.android.coremethods.data.remote.request.CardTokenBodyRequest
-import com.mercadopago.sdk.android.coremethods.data.remote.request.ThreeDSDeviceDataRequest
 import com.mercadopago.sdk.android.coremethods.data.remote.response.CardIssuerResponse
 import com.mercadopago.sdk.android.coremethods.data.remote.response.CardTokenResponse
 import com.mercadopago.sdk.android.coremethods.data.remote.response.IdentificationTypesResponse
@@ -49,9 +48,4 @@ internal interface CoreMethodsService {
         @Query("product_id") productId: String? = PRODUCT_ID,
         @Query("bin") bin: Int?,
     ): Response<List<PaymentMethodResponse>>
-
-    @POST("$BRICKS_API/$VERSION/challenges/threeds/device")
-    suspend fun saveThreeDSDeviceData(
-        @Body request: ThreeDSDeviceDataRequest,
-    ): Response<Unit>
 }
