+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/data/repository/ThreeDSRepositoryImpl.kt
@@ -5,6 +5,7 @@ import com.mercadopago.sdk.android.coremethods.data.remote.mappers.toRequest
 import com.mercadopago.sdk.android.coremethods.domain.model.ResultError
 import com.mercadopago.sdk.android.coremethods.domain.model.ThreeDSChallengeAuthentication
 import com.mercadopago.sdk.android.coremethods.domain.model.params.AuthenticateThreeDSChallengeParams
+import com.mercadopago.sdk.android.coremethods.domain.model.params.SaveThreeDSDeviceDataParams
 import com.mercadopago.sdk.android.coremethods.domain.model.params.UpdateThreeDSChallengeStatusParams
 import com.mercadopago.sdk.android.coremethods.domain.repository.ThreeDSRepository
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
@@ -23,4 +24,10 @@ internal class ThreeDSRepositoryImpl(
             challengeId = params.challengeId,
             request = params.toRequest(),
         )
+
+    override suspend fun saveThreeDSDeviceData(
+        params: SaveThreeDSDeviceDataParams,
+    ): Result<Unit, ResultError> {
+        return dataSource.saveThreeDSDeviceData(params.toRequest())
+    }
 }
