+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/domain/repository/ThreeDSRepository.kt
@@ -3,6 +3,7 @@ package com.mercadopago.sdk.android.coremethods.domain.repository
 import com.mercadopago.sdk.android.coremethods.domain.model.ResultError
 import com.mercadopago.sdk.android.coremethods.domain.model.ThreeDSChallengeAuthentication
 import com.mercadopago.sdk.android.coremethods.domain.model.params.AuthenticateThreeDSChallengeParams
+import com.mercadopago.sdk.android.coremethods.domain.model.params.SaveThreeDSDeviceDataParams
 import com.mercadopago.sdk.android.coremethods.domain.model.params.UpdateThreeDSChallengeStatusParams
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
 
@@ -13,5 +14,9 @@ internal interface ThreeDSRepository {
 
     suspend fun updateThreeDSChallengeStatus(
         params: UpdateThreeDSChallengeStatusParams,
+    ): Result<Unit, ResultError>
+
+    suspend fun saveThreeDSDeviceData(
+        params: SaveThreeDSDeviceDataParams,
     ): Result<Unit, ResultError>
 }
