+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/domain/usecase/SaveThreeDSDeviceDataUseCase.kt
@@ -4,7 +4,7 @@ import com.mercadopago.sdk.android.coremethods.domain.model.ResultError
 import com.mercadopago.sdk.android.coremethods.domain.model.params.DeviceRenderOptionsParams
 import com.mercadopago.sdk.android.coremethods.domain.model.params.EphemeralPublicKeyParams
 import com.mercadopago.sdk.android.coremethods.domain.model.params.SaveThreeDSDeviceDataParams
-import com.mercadopago.sdk.android.coremethods.domain.repository.CoreMethodsRepository
+import com.mercadopago.sdk.android.coremethods.domain.repository.ThreeDSRepository
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
 
 /**
@@ -13,7 +13,7 @@ import com.mercadopago.sdk.android.coremethods.domain.utils.Result
  * This use case validates the input parameters and delegates the API call to the repository.
  */
 internal class SaveThreeDSDeviceDataUseCase(
-    private val repository: CoreMethodsRepository,
+    private val repository: ThreeDSRepository,
 ) {
     @Suppress("LongParameterList")
     suspend operator fun invoke(
